+++ extra/mariabackup/backup_mysql.cc
@@ -1386,7 +1386,7 @@ operator<<(std::ostream& s, const escape_and_quote& eq)
 	s << '\'';
 	size_t len = strlen(eq.str);
 	char* escaped = (char *)alloca(2 * len + 1);
-	len = mysql_real_escape_string(eq.mysql, escaped, eq.str, len);
+	len = mysql_real_escape_string(eq.mysql, escaped, eq.str, (ulong)len);
 	s << std::string(escaped, len);
 	s << '\'';
 	return s;
